+++ ria-weekend/src/demos/linear_gradient_rectangle.rs
@@ -22,8 +22,8 @@ impl Demo for LinearGradientRectangle {
 
         for j in (0..height).rev() {
             for i in 0..width {
-                let u = i as f32 / width as f32;
-                let v = j as f32 / height as f32;
+                let u = i as f64 / width as f64;
+                let v = j as f64 / height as f64;
                 let ray = Ray::new(origin, lower_left_corner + horizontal * u + vertical * v);
 
                 let c = color(ray);
