+++ ria-weekend/src/types/hitable.rs
@@ -1,13 +1,13 @@
 use crate::types::{Ray, Vec3};
 
 pub struct HitRecord {
-    pub t: f32,
+    pub t: f64,
     pub p: Vec3,
     pub normal: Vec3,
 }
 
 pub trait Hitable {
-    fn hit(&self, _ray: &Ray, _t_min: f32, _t_max: f32) -> Option<HitRecord> {
+    fn hit(&self, _ray: &Ray, _t_min: f64, _t_max: f64) -> Option<HitRecord> {
         None
     }
 }
